+++ src/managednetworkfabric/azext_managednetworkfabric/aaz/latest/networkfabric/fabric/_show.py
@@ -270,6 +270,7 @@ class Show(AAZCommand):
             properties.network_fabric_controller_id = AAZStrType(
                 serialized_name="networkFabricControllerId",
                 flags={"required": True},
+                nullable=True,
             )
             properties.network_fabric_sku = AAZStrType(
                 serialized_name="networkFabricSku",
@@ -314,7 +315,9 @@ class Show(AAZCommand):
             active_commit_batches.Element = AAZStrType()
 
             control_plane_acls = cls._schema_on_200.properties.control_plane_acls
-            control_plane_acls.Element = AAZStrType()
+            control_plane_acls.Element = AAZStrType(
+                nullable=True,
+            )
 
             fabric_locks = cls._schema_on_200.properties.fabric_locks
             fabric_locks.Element = AAZObjectType()
@@ -370,6 +373,7 @@ class Show(AAZCommand):
             storage_account_configuration = cls._schema_on_200.properties.storage_account_configuration
             storage_account_configuration.storage_account_id = AAZStrType(
                 serialized_name="storageAccountId",
+                nullable=True,
             )
             storage_account_configuration.storage_account_identity = AAZObjectType(
                 serialized_name="storageAccountIdentity",
@@ -382,6 +386,7 @@ class Show(AAZCommand):
             )
             storage_account_identity.user_assigned_identity_resource_id = AAZStrType(
                 serialized_name="userAssignedIdentityResourceId",
+                nullable=True,
             )
 
             terminal_server_configuration = cls._schema_on_200.properties.terminal_server_configuration
@@ -414,7 +419,9 @@ class Show(AAZCommand):
             )
 
             trusted_ip_prefixes = cls._schema_on_200.properties.trusted_ip_prefixes
-            trusted_ip_prefixes.Element = AAZStrType()
+            trusted_ip_prefixes.Element = AAZStrType(
+                nullable=True,
+            )
 
             unique_rd_configuration = cls._schema_on_200.properties.unique_rd_configuration
             unique_rd_configuration.nni_derived_unique_rd_configuration_state = AAZStrType(
@@ -481,6 +488,7 @@ class _ShowHelper:
         )
         vpn_configuration_properties_read.network_to_network_interconnect_id = AAZStrType(
             serialized_name="networkToNetworkInterconnectId",
+            nullable=True,
         )
         vpn_configuration_properties_read.option_a_properties = AAZObjectType(
             serialized_name="optionAProperties",