+++ src/managednetworkfabric/azext_managednetworkfabric/aaz/latest/networkfabric/taprule/_update.py
@@ -246,6 +246,7 @@ class Update(AAZCommand):
         _element.destination_id = AAZResourceIdArg(
             options=["destination-id"],
             help="Destination Id. The ARM resource Id may be either Network To Network Interconnect or NeighborGroup.",
+            nullable=True,
         )
         _element.is_timestamp_enabled = AAZStrArg(
             options=["is-timestamp-enabled"],
@@ -589,7 +590,7 @@ class Update(AAZCommand):
 
             _elements = _builder.get(".properties.matchConfigurations[].actions[]")
             if _elements is not None:
-                _elements.set_prop("destinationId", AAZStrType, ".destination_id")
+                _elements.set_prop("destinationId", AAZStrType, ".destination_id", typ_kwargs={"nullable": True})
                 _elements.set_prop("isTimestampEnabled", AAZStrType, ".is_timestamp_enabled")
                 _elements.set_prop("matchConfigurationName", AAZStrType, ".match_configuration_name")
                 _elements.set_prop("truncate", AAZStrType, ".truncate")
@@ -834,6 +835,7 @@ class Update(AAZCommand):
             _element = cls._schema_on_200.properties.match_configurations.Element.actions.Element
             _element.destination_id = AAZStrType(
                 serialized_name="destinationId",
+                nullable=True,
             )
             _element.is_timestamp_enabled = AAZStrType(
                 serialized_name="isTimestampEnabled",